ElderCatcher/requirements.txt
kivy>=2.1.0
google-generativeai>=0.3.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.14

